Refactor cache invalidation logic and consolidate imports
Moved cache invalidation functions from `cache.py` to `cache_managers.py` to align with their relevant contexts and responsibilities. Updated imports across the codebase to reflect these changes, ensuring a cleaner and more maintainable structure.
