+++ common/utils/cache.py
@@ -1,13 +1,11 @@
 # common/utils/cache.py
 import json
-from typing import Union, Optional
-
-import redis
-import hashlib
 import logging
-from functools import wraps
+import hashlib
+import redis
+from typing import Union, Optional, Any, Dict
+
 from common.config import REDIS_URL
-from common.utils.cache_managers import UserCacheManager, SubscriptionCacheManager, FavoriteCacheManager
 
 logger = logging.getLogger(__name__)
 
@@ -62,50 +60,12 @@ def get_entity_cache_key(entity_type: str, entity_id: Union[int, str], suffix: O
     return key
 
 
-def invalidate_user_caches(user_id: int) -> int:
-    """
-    Invalidate all caches related to a user.
-
-    Args:
-        user_id: Database user ID
-
-    Returns:
-        Number of invalidated cache keys
-    """
-    return UserCacheManager.invalidate_all(user_id)
-
-
-def invalidate_subscription_caches(user_id: int, subscription_id: Optional[int] = None) -> int:
-    """
-    Invalidate all subscription-related caches for a user.
-
-    Args:
-        user_id: Database user ID
-        subscription_id: Optional specific subscription ID
-
-    Returns:
-        Number of invalidated cache keys
-    """
-    return SubscriptionCacheManager.invalidate_all(user_id, subscription_id)
-
-
-def invalidate_favorite_caches(user_id: int) -> int:
-    """
-    Invalidate favorite-related caches for a user.
-
-    Args:
-        user_id: Database user ID
-
-    Returns:
-        Number of invalidated cache keys
-    """
-    return FavoriteCacheManager.invalidate_all(user_id)
-
-
 def redis_cache(prefix, ttl=CacheTTL.STANDARD):
     """Cache decorator that uses Redis with standardized TTL"""
 
     def decorator(func):
+        from functools import wraps
+
         @wraps(func)
         def wrapper(*args, **kwargs):
             # Generate cache key
@@ -128,8 +88,21 @@ def redis_cache(prefix, ttl=CacheTTL.STANDARD):
             return result
 
         # Add cache invalidation method to the function
-        wrapper.invalidate_cache = lambda *args, **kwargs: invalidate_cache(prefix, *args, **kwargs)
+        def invalidate_cache(*args, **kwargs):
+            if args or kwargs:
+                # Invalidate specific key
+                key = cache_key(prefix, *args, **kwargs)
+                redis_client.delete(key)
+                logger.debug(f"Invalidated cache key: {key}")
+            else:
+                # Invalidate all keys with this prefix
+                pattern = f"{prefix}:*"
+                keys = redis_client.keys(pattern)
+                if keys:
+                    redis_client.delete(*keys)
+                    logger.debug(f"Invalidated {len(keys)} cache keys with pattern: {pattern}")
 
+        wrapper.invalidate_cache = invalidate_cache
         return wrapper
 
     return decorator
@@ -139,6 +112,8 @@ def async_redis_cache(prefix, ttl=CacheTTL.STANDARD):
     """Cache decorator for async functions"""
 
     def decorator(func):
+        from functools import wraps
+
         @wraps(func)
         async def wrapper(*args, **kwargs):
             # Generate cache key
@@ -161,60 +136,24 @@ def async_redis_cache(prefix, ttl=CacheTTL.STANDARD):
             return result
 
         # Add cache invalidation method to the function
-        wrapper.invalidate_cache = lambda *args, **kwargs: invalidate_cache(prefix, *args, **kwargs)
+        def invalidate_cache(*args, **kwargs):
+            if args or kwargs:
+                # Invalidate specific key
+                key = cache_key(prefix, *args, **kwargs)
+                redis_client.delete(key)
+                logger.debug(f"Invalidated cache key: {key}")
+            else:
+                # Invalidate all keys with this prefix
+                pattern = f"{prefix}:*"
+                keys = redis_client.keys(pattern)
+                if keys:
+                    redis_client.delete(*keys)
+                    logger.debug(f"Invalidated {len(keys)} cache keys with pattern: {pattern}")
 
+        wrapper.invalidate_cache = invalidate_cache
         return wrapper
 
     return decorator
-
-
-def invalidate_cache(prefix, *args, **kwargs):
-    """
-    Invalidate a specific cache entry or pattern of entries
-
-    Args:
-        prefix: Cache prefix to invalidate
-        *args, **kwargs: If provided, invalidates specific key matching these args
-                        If not provided, invalidates all keys with this prefix
-    """
-    if args or kwargs:
-        # Invalidate specific key
-        key = cache_key(prefix, *args, **kwargs)
-        redis_client.delete(key)
-        logger.debug(f"Invalidated cache key: {key}")
-    else:
-        # Invalidate all keys with this prefix
-        pattern = f"{prefix}:*"
-        keys = redis_client.keys(pattern)
-        if keys:
-            redis_client.delete(*keys)
-            logger.debug(f"Invalidated {len(keys)} cache keys with pattern: {pattern}")
-
-
-def cache_warm(prefix, ttl=CacheTTL.STANDARD, data_generator=None):
-    """
-    Warms up cache with provided data or generator function
-
-    Args:
-        prefix: Cache prefix
-        ttl: Cache TTL in seconds
-        data_generator: Function that returns a dict of {key: value} pairs to cache,
-                       where key is everything after the prefix in the cache key
-    """
-    if not data_generator:
-        logger.warning("No data generator provided for cache warming")
-        return
-
-    if callable(data_generator):
-        data = data_generator()
-    else:
-        data = data_generator
-
-    for key_suffix, value in data.items():
-        full_key = f"{prefix}:{key_suffix}"
-        redis_client.set(full_key, json.dumps(value), ex=ttl)
-
-    logger.info(f"Warmed up {len(data)} cache entries with prefix {prefix}")
 
 
 def get_cached(key, default=None):
